feat: added basic scores viewing

+++ app/src/main/java/com/example/theexplorer/services/QRCode.java
@@ -16,7 +16,7 @@ public class QRCode {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
@@ -24,7 +24,7 @@ public class QRCode {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
